+++ waspc/ChangeLog.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## Unreleased
+
+### ⚠️ Breaking Changes
+
+- Wasp now requires Node.js version to be >=v22.12. [#2915](https://github.com/wasp-lang/wasp/pull/2915)
+
+### 🎉 New Features
+
+- New command: `wasp build start`. Lets you run your built Wasp app locally, just like in production. If required environment variables are missing, the command will show errors—helping you confirm which env vars you'll need to set in production. [#2796](https://github.com/wasp-lang/wasp/pull/2796)
+
+## 0.17.1
+
+### 🐞 Bug fixes
+
+- Fixes parsing of `app.client` and `app.server` options in `main.wasp.ts` ([#2999](https://github.com/wasp-lang/wasp/pull/2999))
+
 ## 0.17.0
 
 ### ⚠️ Breaking Changes
@@ -1919,5 +1935,3 @@ For exact details about new syntax, check https://wasp.sh/docs/language/syntax .
 - Added delay on recompilation to avoid redundant recompiling.
 - Added `onAuthSucceededRedirectTo` field in `app`.
 - and more!
-
-## Unreleased changes